docker_2/requirements.txt
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0